==> Datasets/Heckmann_et_al_2017/1_QaQc_curated_ACi.R <==
## This code is used to do the step 1 of the data curation process.
## It is used to analyse and check the data quality. Bad points are removed as well as bad curves.

# Load the 'here' package to easily reference files and folders path, which is robust and independent on your platform (Linux, Windows..)
library(here)

# Find the path of the top relative directory
path=here()

# Set the working directory to the 'Heckmann_et_al_2017' folder where the data is located
setwd(file.path(path,'/Datasets/Heckmann_et_al_2017'))

# Load the ACi data that were processed in step 0
load('0_curated_data.Rdata',verbose=TRUE)

# Adding a Quality Check column with two values: "ok" or "bad", where "bad" correspond to bad data 
# that is not used to fit the ACi curves in the next step.
curated_data$QC='ok'

# Flagging the below 0 Ci as bad data
hist(curated_data$Ci)
curated_data[curated_data$Ci<0,'QC']='bad'

# Checking if there are duplicated CO2r points in the ACi curves and only keeping the first one
Remove_Aci_dupplicate<-function(Aci_curve){
  ##Rounding the ci values
  Aci_curve=round(Aci_curve,-1)
  Remove=rep('NO',length(Aci_curve))
  ## We Remove the last point of the curve if it is a duplicate
  if(length(Aci_curve)%in%which(duplicated(Aci_curve))){Remove[length(Aci_curve)]='YES'}
  ## We now remove the duplicates in the middle of the ACi curves, except the last
  Remove[which(duplicated(Aci_curve[1:(length(Aci_curve)-1)],fromLast = TRUE))]='YES'
  Remove[1:5]='NO'
  Remove[which(duplicated(Aci_curve[1:5],fromLast = TRUE))]='YES'
  return(Remove)
}

#Remove=ave(round(curated_data$CO2r,-1),curated_data$SampleID,FUN=duplicated)
Remove=ave(curated_data$CO2r,curated_data$SampleID,FUN=Remove_Aci_dupplicate)
curated_data[which(Remove=='YES'),'QC']='bad'

# I create a "Quality check" table where I manually inform bad points. 
# When I start the quality check the QC table is empty: QC_table=cbind.data.frame(SampleID_num=c(),Record=c()) 
# I plot the data in the next step (line starting with pdf below). 
# I then manually check the plots and write down the SampleID_num and the Record corresponding to bad points
# This works well if there are few points to remove.
# You can use another method if you need.

QC_table=cbind.data.frame(SampleID_num=c(56,119,119,119,210,213,258),
                          Record=c(11,10,11,12,9,9,1)) 
# Here I flag the bad curves by writing down the SampleID_num of the bad curves.
ls_bad_curve=c(19,79:82,87,96,138,140:142,151,163,164,174,179,195,196,198:208,211,216,217,221,226)
ls_bad_curve=c(ls_bad_curve,12,132,139,147,166,172,176,185,197,209,228) # I removed these curves based on the fittings
# Here I flag all the bad curves and bad points
curated_data[paste(curated_data$SampleID_num,curated_data$Record)%in%paste(QC_table$SampleID_num,QC_table$Record),'QC']='bad'
curated_data[curated_data$SampleID_num%in%ls_bad_curve,'QC']='bad'


# Creating a PDF file where all the ACi curves are plot. The points in red are bad points or curves and the points in black are ok.
pdf(file='1_QA_QC_Aci.pdf',)
for(SampleID_num in sort(unique(curated_data$SampleID_num))){
plot(x=curated_data[curated_data$SampleID_num==SampleID_num,'Ci'],y=curated_data[curated_data$SampleID_num==SampleID_num,'A'],main=unique(curated_data[curated_data$SampleID_num==SampleID_num,'SampleID_num']),cex=2,xlab='Ci',ylab='A')
  text(x=curated_data[curated_data$SampleID_num==SampleID_num,'Ci'],y=curated_data[curated_data$SampleID_num==SampleID_num,'A'], labels=curated_data[curated_data$SampleID_num==SampleID_num,'Record'],cex=0.7)
if(any(curated_data[curated_data$SampleID_num==SampleID_num,'QC']=='bad')){
  points(x=curated_data[curated_data$SampleID_num==SampleID_num&curated_data$QC=='bad','Ci'],y=curated_data[curated_data$SampleID_num==SampleID_num&curated_data$QC=='bad','A'],cex=2,col='red')
  text(x=curated_data[curated_data$SampleID_num==SampleID_num&curated_data$QC=='bad','Ci'],y=curated_data[curated_data$SampleID_num==SampleID_num&curated_data$QC=='bad','A'], labels=curated_data[curated_data$SampleID_num==SampleID_num&curated_data$QC=='bad','Record'],cex=0.7,col='red')
}
}
dev.off()

# Keeping only the good points and curves
curated_data=curated_data[curated_data$QC=='ok',]

# Saving the Quality checked data to be used in the next step.
save(curated_data,file='1_QC_ACi_data.Rdata')
